Add and rearrange checked and clear functions
