+++ src/modules/Check$Clear.js
@@ -19,4 +19,26 @@ const checked = (list) => {
   });
 };
 
-export { checked as default };
+const rearrangeIndex = (list) => {
+  const boxes = document.querySelectorAll('.inputTasks');
+  for (let i = 1; i <= list.taskList.length; i += 1) {
+    list.taskList[i - 1].index = i;
+    boxes[i - 1].dataset.value = `${i}`;
+  }
+  list.index = list.taskList.length + 1;
+  populateStorage(list);
+};
+
+const clear = (list) => {
+  const clear = document.querySelector('#clear');
+  clear.addEventListener('click', () => {
+    const checked = document.querySelectorAll('.checked');
+    checked.forEach((item) => {
+      item.parentNode.parentNode.removeChild(item.parentNode);
+    });
+    list.taskList = list.taskList.filter((item) => item.completed !== true);
+    rearrangeIndex(list);
+  });
+};
+
+export { checked, clear, rearrangeIndex };
